+++ lib/ssm_storage/db.rb
