+++ lib/ssm_storage/empty.rb
